requirements.txt

tavily-python
pandas

langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph
python-dotenv 
requests
beautifulsoup4
playwright
python-telegram-bot 
docling
pyarrow
ipywidgets
spacy
langdetect
tavily-python
